# Conflicts:
#	.ci/configure-libevent.sh
#	.gitlab-ci.yml
#	gframe/data_manager.cpp
#	gframe/data_manager.h
#	gframe/deck_manager.h
#	gframe/game.cpp
#	gframe/game.h
#	gframe/netserver.h
#	gframe/premake5.lua
#	gframe/replay.h
#	premake5.lua

+++ gframe/deck_manager.h
@@ -70,7 +70,7 @@ public:
 	const wchar_t* GetLFListName(unsigned int lfhash);
 #endif
 	const LFList* GetLFList(unsigned int lfhash);
-	unsigned int CheckDeck(const Deck& deck, unsigned int lfhash, int rule);
+	uint32_t CheckDeck(const Deck& deck, unsigned int lfhash, size_t rule);
 #ifndef YGOPRO_SERVER_MODE
 	bool LoadCurrentDeck(const wchar_t* file, bool is_packlist = false);
 	bool LoadCurrentDeck(int category_index, const wchar_t* category_name, const wchar_t* deckname);
